Load env vars immediately before executing a command
This is to avoid collisions between the env vars set from the config
vars and the env vars used in the script (both might overwrite the
others). In the present way, the risk for collisions is localised and
minimised: all the env vars of the script that could clash with the env
vars from the config vars are prefixed with BUILDPACK_RUN_, so if there
is no config var with this prefix, there won't be any collisions.